verocta-ai-backend/backend/requirements.txt
# VeroctaAI Financial Intelligence Platform - Production Requirements
# Web framework
flask==3.1.1
werkzeug==3.1.3

# CORS and JWT authentication
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.1

# Data processing
pandas==2.3.1
numpy>=1.26.0

# AI integration
openai==1.98.0

# PDF generation
reportlab==4.4.3
matplotlib>=3.5.0

# Image processing
Pillow>=10.0.0

# Production server
gunicorn==23.0.0

# Date utilities
python-dateutil>=2.8.2

# Database
psycopg2-binary==2.9.9

# Additional utilities
requests==2.31.0

# Payment processing
stripe==7.8.0

# Email services
resend==0.8.0

# Google Sheets integration
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2

# Supabase client
supabase==2.3.0
bcrypt
email-validator
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
google-api-python-client
google-auth
google-auth-oauthlib
gunicorn
matplotlib
numpy
openai
pandas
pillow
psycopg2-binary
python-dateutil
python-dotenv
reportlab
requests
resend
sqlalchemy
stripe
supabase
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
bcrypt==4.1.2
email-validator
flask==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gunicorn==23.0.0
matplotlib>=3.5.0
numpy>=1.26.0
openai==1.98.0
pandas==2.3.1
Pillow>=10.0.0
psycopg2-binary==2.9.9
python-dateutil>=2.8.2
python-dotenv==1.0.1
reportlab==4.4.3
requests==2.31.0
resend==0.8.0
stripe==7.8.0
supabase==2.3.0
werkzeug==3.1.3
